Move clusterconfig
